requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.20
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
